backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-tenants==3.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
pip==25.0.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.7.0